+++ mathFunctions.hpp
@@ -1,9 +1,9 @@
-#include <gmpxx.h>
+/*#include <gmpxx.h>
 
 mpz_class fatorial(mpz_class num);
 int menosUmElevadoAn(int n);
-
+*/
 
 int fatorialINT(int i);
 int menosUmElevadoAnINT(int i);
-int potenciaINT(int base, int pot);
+float potenciaINT(float base, int pot);
